Test problems with conversion precision
